# Eliminate use of external commands from ocamltest - take 2