--- a/Changes
+++ b/Changes
@@ -178,6 +178,10 @@ Working version
   table format and generic hash function that were in use before OCaml 4.00.
   (Xavier Leroy, review by Nicolás Ojeda Bär)
 
+- #9797: Add Sys.mkdir.
+  (David Allsopp, review by Nicolás Ojeda Bär, Sébastien Hinderer and
+   Xavier Leroy)
+
 ### Other libraries:
 
 * #9206, #9419: update documentation of the threads library;