--- a/ocamltest/main.ml
+++ b/ocamltest/main.ml
@@ -152,9 +152,9 @@ let test_file test_filename =
   let test_build_directory_prefix =
     get_test_build_directory_prefix test_directory in
   let clean_test_build_directory () =
-    ignore
-      (Sys.command
-         (Filename.quote_command "rm" ["-rf"; test_build_directory_prefix]))
+    try
+      Sys.rm_rf test_build_directory_prefix
+    with Sys_error _ -> ()
   in
   clean_test_build_directory ();
   Sys.make_directory test_build_directory_prefix;