--- a/ocamltest/ocamltest_stdlib.ml
+++ b/ocamltest/ocamltest_stdlib.ml
@@ -15,6 +15,8 @@
 
 (* A few extensions to OCaml's standard library *)
 
+module Unix = Ocamltest_unix
+
 (* Pervaisive *)
 
 let input_line_opt ic =
@@ -164,8 +166,6 @@ module Sys = struct
 
   let force_remove file =
     if file_exists file then remove file
-
-  external has_symlink : unit -> bool = "caml_has_symlink"
 
   let with_chdir path f =
     let oldcwd = Sys.getcwd () in