--- a/ocamltest/ocamltest_stdlib.ml
+++ b/ocamltest/ocamltest_stdlib.ml
@@ -105,15 +105,6 @@ module Sys = struct
       with Sys_error err ->
         raise (Sys_error (Printf.sprintf "Failed to remove %S (%s)" path err))
 
-  let run_system_command prog args =
-    let command = Filename.quote_command prog args in
-    match Sys.command command with
-    | 0 -> ()
-    | _ as exitcode ->
-      Printf.eprintf "System command %s failed with status %d\n%!"
-        command exitcode;
-      exit 3
-
   let rec make_directory dir =
     if Sys.file_exists dir then ()
     else let () = make_directory (Filename.dirname dir) in