--- a/ocamltest/ocamltest_stdlib.mli
+++ b/ocamltest/ocamltest_stdlib.mli
@@ -48,6 +48,7 @@ module Sys : sig
   val file_is_empty : string -> bool
   val run_system_command : string -> string list -> unit
   val make_directory : string -> unit
+  val rm_rf : string -> unit
   val string_of_file : string -> string
   val iter_lines_of_file : (string -> unit) -> string -> unit
   val dump_file : out_channel -> ?prefix:string -> string -> unit