--- a/ocamltest/ocamltest_unix.mli
+++ b/ocamltest/ocamltest_unix.mli
@@ -16,4 +16,5 @@
     temptation to use the Unix module directly in ocamltest. *)
 
 val has_symlink : unit -> bool
+val symlink : ?to_dir:bool -> string -> string -> unit
 val chmod : string -> int -> unit