--- a/ocamltest/ocamltest_unix_dummy.ml
+++ b/ocamltest/ocamltest_unix_dummy.ml
@@ -1,0 +1,16 @@
+(**************************************************************************)
+(*                                                                        *)
+(*                                 OCaml                                  *)
+(*                                                                        *)
+(*                 David Allsopp, OCaml Labs, Cambridge.                  *)
+(*                                                                        *)
+(*   Copyright 2020 David Allsopp Ltd.                                    *)
+(*                                                                        *)
+(*   All rights reserved.  This file is distributed under the terms of    *)
+(*   the GNU Lesser General Public License version 2.1, with the          *)
+(*   special exception on linking described in the file LICENSE.          *)
+(*                                                                        *)
+(**************************************************************************)
+
+(* Dummy implementations for when the Unix library isn't built *)
+let has_symlink () = false