--- a/ocamltest/ocamltest_unix_dummy.ml
+++ b/ocamltest/ocamltest_unix_dummy.ml
@@ -14,4 +14,5 @@
 
 (* Dummy implementations for when the Unix library isn't built *)
 let has_symlink () = false
+let symlink ?to_dir:_ _ _ = invalid_arg "symlink not available"
 let chmod _ _ = invalid_arg "chmod not available"