--- a/ocamltest/ocamltest_unix_real.ml
+++ b/ocamltest/ocamltest_unix_real.ml
@@ -25,4 +25,5 @@ let wrap f x =
     in
       raise (Sys_error msg)
 
+let symlink ?to_dir source = wrap (Unix.symlink ?to_dir source)
 let chmod file = wrap (Unix.chmod file)