--- a/otherlibs/unix/mkdir.c
+++ b/otherlibs/unix/mkdir.c
@@ -13,9 +13,15 @@
 /*                                                                        */
 /**************************************************************************/
 
+#ifndef _WIN32
 #include <sys/types.h>
 #include <sys/stat.h>
+#endif
+
+#define CAML_INTERNALS
 #include <caml/mlvalues.h>
+#include <caml/osdeps.h>
+#include <caml/misc.h>
 #include <caml/memory.h>
 #include <caml/signals.h>
 #include "unixsupport.h"
@@ -23,12 +29,12 @@
 CAMLprim value unix_mkdir(value path, value perm)
 {
   CAMLparam2(path, perm);
-  char * p;
+  char_os * p;
   int ret;
   caml_unix_check_path(path, "mkdir");
-  p = caml_stat_strdup(String_val(path));
+  p = caml_stat_strdup_to_os(String_val(path));
   caml_enter_blocking_section();
-  ret = mkdir(p, Int_val(perm));
+  ret = mkdir_os(p, Int_val(perm));
   caml_leave_blocking_section();
   caml_stat_free(p);
   if (ret == -1) uerror("mkdir", path);