--- a/stdlib/sys.mli
+++ b/stdlib/sys.mli
@@ -94,6 +94,12 @@ external time : unit -> (float [@unboxed]) =
 external chdir : string -> unit = "caml_sys_chdir"
 (** Change the current working directory of the process. *)
 
+external mkdir : string -> int -> unit = "caml_sys_mkdir"
+(** Create a directory with the given permissions.
+
+    @since 4.12.0
+*)
+
 external getcwd : unit -> string = "caml_sys_getcwd"
 (** Return the current working directory of the process. *)