… otherlibs/win32unix/mkdir.c