This reworks #9787 with only `Sys.mkdir` and `Sys.rmdir` added to the standard library. With just these two new primitives, it is possible to build an `ocamltest` which doesn't use external commands (it will therefore copy sources on all platforms, since it lacks `Unix.symlink`). Even better, the `MIN_BUILD` run on Travis will continuously check this, since it doesn't build the Unix library. On Windows, if ocamltest is built without the Unix library and a test sets the readonly attribute on a file then ocamltest will be unable to remove it. We can live with this since, unlike Dune, we're in total control of what the users of ocamltest do!

If the Unix library is built, ocamltest will link with it. This eliminates the need for its duplicate `caml_has_symlink` stub (it just uses `Unix.has_symlink` and, clearly, if the Unix library hasn't been built then we know the outcome of the `has_symlink` predicate) and it's then easy to use `Unix.symlink` instead of `ln -sf`. In this instance, Windows may use `Unix.chmod` in order to remove the read-only attribute from files it is trying to delete.

In order to guard against accidental use of `Unix` functions, I have altered ocamltest so that all modules open `Ocamltest_stdlib` (and therefore get `Ocamltest_stdlib.Unix` instead of the Unix library's `Unix`).

cc @xavierleroy, @shindere, @nojb