@nojb 2020-07-23
Looks nice! However, I for review purposes I feel it would be simpler to separate the additions to `Sys` in a different PR (since they can stand on their own). What do you think?