@shindere 2020-07-23
Nicolás Ojeda Bär (2020/07/23 07:35 -0700):
> Looks nice! However, I for review purposes I feel it would be simpler
> to separate the additions to `Sys` in a different PR (since they can
> stand on their own). What do you think?

Yes, I agree with this.