@shindere 2020-07-23
David Allsopp (2020/07/23 07:37 -0700):
> To be honest, I'd prefer to get on with other branches than re-stack
> this PR again - the changes to `Sys` are now very small.

Well the concerns seem different enough to me to deserve distincts PRs.