@xavierleroy 2020-07-23
> Well the concerns seem different enough to me to deserve distincts PRs.

I don't mind and can review in either case (one or two PRs), but please let's make up our collective mind, and then ping me.