@dra27 2020-07-23
A review of this would be much appreciated! As I said in the previous PR, I’m very happy to blow away the commits altering `sys.mlp` and `sys.mli` if there needs to be discussion which might lead to those functions _not_ being exposed, but the duplicating of those primitives in the runtime is intimately tied to this PR, as *requiring* Unix to build ocamltest involves some CI golf with the MIN_BUILD test and also changes to `configure.ac` which I’d prefer not to have to do.