@dra27 2020-07-23
> Error from Travis:
> 
> > File "ocamltest_unix_dummy.ml", line 17, characters 14-20:
> > Error (warning 27 [unused-var-strict]): unused variable to_dir.

I can never remember the syntax for avoiding that one! <sigh>