@dra27 2020-07-24
Yeah, I expected the `Ocamltest_stdlib` thing to be easier but, as usual, statically generated `ocamldep` dependencies got in the way :trollface: In its defence, the manipulation rules for `$(open_flag)` are as in `runtime/Makefile` and it was at least possible to get the modules listed in one place. The alternative might be to ensure that all the .ml files have `open Ocamltest_stdlib` at the top and suppress the unused `open` warning?