@dra27 2020-07-24
An obvious alternative to the `-open Ocamltest_config` is only to do `-I ../otherlibs/unix` (or `-I ../otherlibs/win32unix`) when compiling the `ocamltest_unix` module and when linking the final executable.