@dra27 2020-07-24
@shindere (accidentally in 7979) writes:
> I just reviewed the whole code. Many thanks for this hard work, @dra27.

My pleasure!
> It's really great that we don't have to use external commands anymore.

I was suspicious at slower runs on AppVeyor, which I expected were just down to load on their servers at the time, but I benchmarked this PR to check. It confirms a performance improvement - I'm waiting for some further runs to finish and I'll post actual numbers here. In particular, I'm checking whether using symlinks is actually noticeable on Windows.
> The addition of `mkdir` and `rmdir` to the stdlib is also a very nice
benefit, I think many users will be happy to have this!
>
> I have only two comments.
>
> 1. I think this PR would definitely deserve a Changes entry.

Indeed - I was adopting the @xavierleroy technique of adding it at the end, also partly because I think will be added in two places (one for the Sys additions and one for the ocamltest changes - I think we've done this with other PRs too).
> 2. I didn't understand why the part about opening `Ocamltest_stdlib`
everywhere was necessary. Even if it may be a good idea (which I'm not sure
about), couldn't it be in
another PR? Or is it really a requirement to implement this PR? why?

The reason as part of this PR was to ensure that the `Unix` library couldn't be used by mistake, so the idea was to ensure `Ocamltest_stdlib` was always opened and so shadowing the `Unix` module. However, I think that controlling when the `-I` flag is given is a better solution.

> Would you mind cleaning the history a bit even while we are reviewing?

I was certainly planning on doing so before merge - happy to do that now if it helps?