@shindere 2020-07-24
David Allsopp (2020/07/24 03:51 -0700):
> I was certainly planning on doing so before merge - happy to do that
> now if it helps?

Well I'm personnally happy with the code so I don't think I'll spend
more time reviewing, except if I'm asked to or if something else gets
added.