@dra27 2020-07-24
I did a simple benchmark on Windows with msvc64/VS2017 averaging 3 runs of the testsuite (after running `make -C testsuite lib tools`):
- 9f25b62 (original base of this PR): testsuite took 38:43
- This PR with `Actions_helpers.setup_symlinks` set always to copy: 37:38
- This PR unmodified (so with symbolic links): 37:35

However, that was with my moderately popular and rather useful clink plugin disabled. With clink enabled, our crippled `Sys.command` (PR to come at some point) takes the testsuite to 50:46 for 9f25b62 and this PR reduces that to 43:03.