@xavierleroy 2020-07-24
Concerning performance, I ran @dra27's benchmark on my Linux workstation:

Time       | Current (9f25b62)  | This PR
-----------|--------------------|------------------
real       | 4m26.8s            | 4m24.6s
user       | 3m27.3s            | 3m24.4s
sys        | 0m25.5s            | 0m26.3s

I was expecting a bigger difference.