@dra27 2020-07-27
Thanks for the review, @xavierleroy - I'll push the changes in just a sec. For the benchmarking, my only concern was that nothing had got unexpectedly slower (I made these changes while speeding up the build on Windows, but these changes are about portability and stability rather than speed). glibc implements system in terms of `posix_spawn`, so I guess it's just demonstrating that process creation is as good as it's cracked up to be on Unix?!