@dra27 2020-07-27
The bootstraps were unnecessary and have gone. I'm also putting this [through precheck](https://ci.inria.fr/ocaml/job/precheck/470/)