@shindere 2020-07-27
David Allsopp (2020/07/27 01:38 -0700):
> @dra27 commented on this pull request.
>
>
>
> >
>    let rec make_directory dir =
>      if Sys.file_exists dir then ()
> -    else (make_directory (Filename.dirname dir); mkdir dir)
> +    else let () = make_directory (Filename.dirname dir) in
> +         if not (Sys.file_exists dir) then
> +           try Sys.mkdir dir 0o777
> +           with _ -> exit 3
>
> @nojb is correct - and catching all errors instead of `Sys_error _` is a typo. An error message might be a good idea too.
>
> Looking at the rest of ocamltest, we have exit code 1 for failed tests and invocation errors (bad command line, corrupt test etc.), exit code 2 for errors in a test itself (mainly used in `Tsl_semantics`) and then exit code 3 for errors in "external" commands, but that now is just this place.
>
> Looking in `Main.tsl_block_of_file_safe`, perhaps this the handling
> could just be removed completely and allow that to display the
> `Sys_error` as it would with any other I/O problem?

That would be fine with me.