@dra27 2020-07-27
There's an flambda problem with this which appeared after the `-open Ocamltest_stdlib` trick was removed - I'm just looking into it.