@dra27 2020-07-27
Turns out `-opaque` is needed when compiling `Ocamltest_unix` in order for the trick of not including `-I ../otherlibs/unix` to work with flambda, which I've added.

Re-running [through precheck](https://ci.inria.fr/ocaml/job/precheck/472/), hopefully with rather fewer red globes this time!